requirements.txt
git+https://github.com/oele3110/modbusReader@main
pyModbusTCP~=0.3.0
paho-mqtt~=2.1.0
aiohttp~=3.11.13
pyyaml~=6.0
fastapi~=0.115.0
uvicorn[standard]~=0.32.0
python-jose[cryptography]~=3.3.0
passlib[bcrypt]~=1.7.4
influxdb-client[async]~=1.45.0